Linear regressions
It seems that linear regressions will be stored in the
correlations_processing doc in the PRIMER folder. This is not ideal but
I'm worried about moving the references.

Lots of changes to linear regression code as I'm writing up Results.
Also made changes to raw Inundation data (copied % inundation for Ref
from LM cores for Carrot and NOAA, since the elevations were roughly the
same and I need complete data sets). Reconfigured some of the
correlation raw data to make it easier to manipulate sediment parameters
in regression equations.