Continuing regression transformations
After talking with Mike about statistical vs pragmatic approach to
regression transformations, played around with some of the code to
compare R2 and p-values for transformed & untransformed data. Continuing
to determine whether it's appropriate to transform. Also created
function to add Age as a numerical value, rather than a character
that'll be converted to a factor, for use in graphing.