Finishing regression transformations
Finished (for now) explorations of transforming data for regressions.
Organized the code to separate checking assumptions and regressing the
transformed data. Ultimately decided (for now) to stick with
untransformed data, as explained in code. Attempted to really separate
chunks in this R Markdown to keep stages of data analysis distinct.