Updated chl-a and thesis ANOVA assumptions
Continuing to work through chl-a data. Realized in the process that some
of my ANOVA assumptions were tested incorrectly during the thesis data
analysis- did not test for normality of residuals. Fixed that also
explored Levene Test with various groupings (still unsure of correct
structure but seems to be ok anyway).